requirements.txt
python-telegram-bot
telegram
numpy 
python-dotenv
networkx 
urllib3
beautifulsoup4
requests
matplotlib
sympy
markov
bs4